chore: centralize setup constants
- Constants belonging to the setup directory structure should be
  placed in setup/constants.rs
- Constants belonging to tools and tests directory structures should be
  placed in tools/constants.rs
